文件管理功能代码上传

+++ src/main/java/com/camelot/pmt/filemanage/service/impl/FileManageGroupServiceImpl.java
@@ -27,6 +27,11 @@ public class FileManageGroupServiceImpl implements FileManageGroupService {
         Long parentId = fileManageGroup.getParentId();
         if(parentId==null){//当添加的文件夹是父节点时
             fileManageGroup.setParentId((long) 0);
+        }else{
+            Long projectId = fileManageGroup.getProjectId();
+            if(projectId!=null){
+                fileManageGroup.setProjectId(null);//如果是子文件夹的时候设置项目id（projectID）为null
+            }
         }
         Long  createId= (Long) request.getSession().getAttribute("  ");//从session获取创建者
         Date date = new Date();
@@ -47,19 +52,29 @@ public class FileManageGroupServiceImpl implements FileManageGroupService {
     public Boolean deleteFileGroup(FileManageGroup fileManageGroup) {
         Long count = null;//删除记录数
         Long id = fileManageGroup.getId();//文件夹id
+        int i = fileManageGroupMapper.deleteByPrimaryKey(id);
     /* List fileManageGroupIds=fileManageGroupMapper.selectFileManagerGroupByParentId(id);//查询id下的所有子文件夹id
         fileManageGroupIds.add(id);//添加当前文件夹id到集合
         fileManageGroupMapper.deleteBatchFileGroupById();//批量删除文件夹
        List fileManagerIds= fileManageMapper.selectFileManagerByGroupId(id);//查询文件夹下的所有文件id
         fileManageMapper.deleteBatchFileById();*/
+        try{
+            deleteFileGroupAndFileById(id);
+            return true;
+        }catch (Exception e){
+            System.out.println("删除失败");
+            return false;
+        }
 
-        // deleteFileGroupAndFileById(id);
-        return null;
     }
 
     @Override
-    public Boolean updateFileGroupById(FileManageGroup fileManageGroup) {
-        int i = fileManageGroupMapper.updateByPrimaryKey(fileManageGroup);//文件夹修改
+    public Boolean updateFileGroupById(HttpServletRequest request,FileManageGroup fileManageGroup) {
+        Date date = new Date();//修改时间
+        Long  modifyUserID = (Long) request.getSession().getAttribute("");
+        fileManageGroup.setModifyTime(date);
+        fileManageGroup.setModifyUserId(modifyUserID);
+        int i = fileManageGroupMapper.updateByPrimaryKeySelective(fileManageGroup);//文件夹修改
         Boolean b=true;
         if(i==0){
             b=false;
@@ -80,15 +95,19 @@ public class FileManageGroupServiceImpl implements FileManageGroupService {
         List<FileManageGroup> groupList=fileManageGroupMapper.selectFileGroup(fileManageGroup);
         return groupList;
     }
- /*public void deleteFileGroupAndFileById(Long id){
+ public void  deleteFileGroupAndFileById(Long id){
         List<Long> fileManageGroupIds=fileManageGroupMapper.selectFileManagerGroupByParentId(id);//获取子文件夹的id
-        fileManageGroupIds.add(id);//添加当前文件夹id到集合
-        fileManageGroupMapper.deleteBatchFileGroupById(fileManageGroupIds);//批量删除文件夹
+     if(fileManageGroupIds.size()>0){
+         fileManageGroupMapper.deleteBatchFileGroupById(fileManageGroupIds);//批量删除文件夹
+     }
         List fileManagerIds= fileManageMapper.selectFileManagerByGroupId(id);//查询文件夹下的所有文件id
-        fileManageMapper.deleteBatchFileById();//删除文件夹下的文档
-        for (Long l:
+     if(fileManagerIds.size()>0){
+         fileManageMapper.deleteBatchFileById(fileManagerIds);//删除文件夹下的文档
+     }
+       for (Long l:
                 fileManageGroupIds) {
             deleteFileGroupAndFileById(l);
         }
-    }*/
+
+    }
 }
